// src/app/shared/Interfaces/characterInterface.ts
export interface Character {
  id: number;
  src: any;
  name: string;
  birth_year: string;
  eye_color: string;
  films: string[];
  gender: string;
  created: string;
  status: string;
  skin_color: string;
  hair_color: string;
  height: string;
  homeworld: string;
  mass: string;
  edited: string;
  url: string;
  species: string[];
  vehicles: string[];
  starships: string[];
}