// src/app/shared/Interfaces/ispecie.ts
export interface ISpecie {
  name: string; //Search Field
  classification: string;
  designation: string;
  average_height: string;
  average_lifespan: string;
  eye_colors: string;
  hair_colors: string;
  skin_colors: string;
  language: string;
  homeworld: string;
  people: string[];
  films: string[];
  url: string[];
  created: string;
  edited: string;
}